improve events list > removed unused markup  and moved HeaderButton
